fix: update QUERY_DEADLINE_HOURS validation to allow non-integer positive numbers

+++ server/src/configs/env.config.js
@@ -43,7 +43,7 @@ const envSchema = z.object({
   VECTOR_SEARCH_INDEX: z.string().default("faq_vector_index"),
   VECTOR_SIMILARITY_THRESHOLD: z.coerce.number().min(0).max(1).default(0.82),
 
-  QUERY_DEADLINE_HOURS: z.coerce.number().int().positive().default(1),
+  QUERY_DEADLINE_HOURS: z.coerce.number().positive().default(1),
   MIN_CONTRIBUTOR_RESPONSES: z.coerce.number().int().positive().default(2),
 
   RATE_LIMIT_WINDOW_MS: z.coerce.number().int().positive().default(900000),
